Move related action hooks into one file
