+++ frontend/awx/administration/instance-groups/InstanceGroupPage/hooks/useIGInstanceRowActions.tsx
@@ -1,8 +1,8 @@
 import { useMemo } from 'react';
 import { IPageAction } from '../../../../../../framework';
 import { Instance } from '../../../../interfaces/Instance';
-import { useRunHealthCheckRowAction } from '../../../instances/hooks/useRunHealthCheckRowAction';
-import { useToggleInstanceRowAction } from '../../../instances/hooks/useToggleInstanceRowAction';
+import { useRunHealthCheckRowAction } from '../../../instances/hooks/useInstanceRowActions';
+import { useToggleInstanceRowAction } from '../../../instances/hooks/useInstanceRowActions';
 
 export function useIGInstanceRowActions(onComplete: (instances: Instance[]) => void) {
   const toggleInstanceRowAction: IPageAction<Instance> = useToggleInstanceRowAction(onComplete);
