+++ frontend/awx/administration/instance-groups/InstanceGroupPage/hooks/useIGInstanceToolbarActions.tsx
@@ -2,7 +2,7 @@ import { useMemo } from 'react';
 import { IPageAction } from '../../../../../../framework';
 import { IAwxView } from '../../../../common/useAwxView';
 import { Instance } from '../../../../interfaces/Instance';
-import { useRunHealthCheckToolbarAction } from '../../../instances/hooks/useRunHealthCheckToolbarAction';
+import { useRunHealthCheckToolbarAction } from '../../../instances/hooks/useInstanceToolbarActions';
 
 export function useIGInstanceToolbarActions(view: IAwxView<Instance>) {
   const healthCheckAction = useRunHealthCheckToolbarAction(view, true);
